backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pgvector
python-multipart
redis
celery
# AI Dependencies
opencv-python-headless
insightface
onnxruntime
numpy
requests
google-auth
stream-zip
Pillow
pillow-heif
